Add typings
